Some files structure & Change theme
