+++ index.js
@@ -3,7 +3,7 @@
  */
 
 import {AppRegistry} from 'react-native';
-import App from './App';
 import {name as appName} from './app.json';
 import SearchItem from './SearchItem'
-AppRegistry.registerComponent(appName, () => App);
+import SearchComp from './Search/SearchComp'
+AppRegistry.registerComponent(appName, () => SearchComp);
